backend/requirements.txt
Flask==2.3.3
apify-client==1.7.1
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
newspaper3k==0.2.8
lxml==4.9.3
lxml_html_clean==0.4.3
Pillow==10.0.1
google-generativeai==0.8.3